requirements.txt
aiogram~=2.14
python-dotenv==0.20.0